Refactor: Split api.py into domain-specific repositories
Extract domain-specific data access logic from the monolithic api.py
(3278 lines) into focused repository classes under repositories/:

- XpBoostRepository (xp_boost_repository.py)
- WarningRepository (warning_repository.py)
- TriggerMessageRepository (trigger_message_repository.py)
- TwitchRepository (twitch_repository.py)
- LogBlacklistRepository (log_blacklist_repository.py)

Each repository follows the established pattern from
LevelConfigRepository and LevelRoleRepository, using deferred
imports from api.py for core DB functions (execute_query,
execute_action, etc.) to avoid circular imports.

Module-level singletons are provided for easy import. Backward-
compatible wrapper functions in api.py delegate to the repos.


Closes #1607
